feat: implement api-utils
