+++ app/api/admin/route.ts
@@ -1,16 +1,11 @@
 import { NextRequest, NextResponse } from 'next/server';
-import { createResponseWithError } from '@/lib/api-utils';
+import { checkAdminAccess, createResponseWithError, getUserAndValidate, responseInit } from '@/lib/api-utils';
 import { ApiRequestError } from '@/lib/error';
-import { checkAdminAccess, getUserAndCheckUser, getUserIdAndCheckAccess } from '@/lib/next-auth/utils';
-import { headers } from '@/lib/request-headers';
 
 export async function GET(req: NextRequest) {
-  const responseInit = { headers };
-
   try {
-    const userId = await getUserIdAndCheckAccess(req);
-    const user = await getUserAndCheckUser(userId);
-    checkAdminAccess(user.role.name);
+    const user = await getUserAndValidate(req);
+    checkAdminAccess(user?.role?.name);
 
     return NextResponse.json(
       {
