+++ app/api/user/route.ts
@@ -1,15 +1,10 @@
 import { NextRequest, NextResponse } from 'next/server';
-import { createResponseWithError } from '@/lib/api-utils';
+import { createResponseWithError, getUserAndValidate, responseInit } from '@/lib/api-utils';
 import { ApiRequestError } from '@/lib/error';
-import { getUserAndCheckUser, getUserIdAndCheckAccess } from '@/lib/next-auth/utils';
-import { headers } from '@/lib/request-headers';
 
 export async function GET(req: NextRequest) {
-  const responseInit = { headers };
-
   try {
-    const userId = await getUserIdAndCheckAccess(req);
-    const user = await getUserAndCheckUser(userId);
+    const user = await getUserAndValidate(req);
 
     return NextResponse.json(
       {
